@@ -21,7 +21,9 @@ class InvalidTokenException extends Exception
     }
 
     /**
-     * @codeCoverageIgnore
+     * Get the exception's context information.
+     *
+     * @return array<string, mixed>
      */
     public function context(): array
     {
