@@ -3,6 +3,6 @@
 declare(strict_types=1);
 
 return [
-    'next' => '下一頁 &raquo;',
+    'next'     => '下一頁 &raquo;',
     'previous' => '&laquo; 上一頁',
 ];
